fix(rocsparse): compute bsrsv solve grid product in 64-bit (AISPARSE-656)
The bsrsv solve/gather launches built their grid as (wfsize * mb - 1) /
bsize + 1 (and the analogous nnzb form) using signed rocsparse_int
arithmetic. At a 64-wide wavefront the wfsize * mb product overflows a
signed 32-bit int at ~33.5M block rows, which is undefined behavior and
yields a corrupt grid dimension. Cast the leading factor to int64_t so
the product and division are performed in 64-bit before narrowing to the
grid dimension, mirroring the existing correct level-2 idioms. The final
grid value still fits in unsigned int.

+++ projects/rocsparse/library/src/level2/rocsparse_bsrsv_solve.cpp
@@ -34,7 +34,7 @@ namespace rocsparse
 {
 #define LAUNCH_BSRSV_GTHR_DIM(bsize, wfsize, dim)                                            \
     RETURN_IF_HIPLAUNCHKERNELGGL_ERROR((rocsparse::bsr_gather<wfsize, bsize / wfsize, dim>), \
-                                       dim3((wfsize * nnzb - 1) / bsize + 1),                \
+                                       dim3((static_cast<int64_t>(wfsize) * nnzb - 1) / bsize + 1), \
                                        dim3(wfsize, bsize / wfsize),                         \
                                        0,                                                    \
                                        stream,                                               \
@@ -95,7 +95,7 @@ namespace rocsparse
 #define LAUNCH_BSRSV_LOWER_SHARED(bsize, wfsize, dim, arch)           \
     RETURN_IF_HIPLAUNCHKERNELGGL_ERROR(                               \
         (bsrsv_lower_shared<bsize, wfsize, dim, arch>),               \
-        dim3((wfsize * mb - 1) / bsize + 1),                          \
+        dim3((static_cast<int64_t>(wfsize) * mb - 1) / bsize + 1),    \
         dim3(bsize),                                                  \
         0,                                                            \
         stream,                                                       \
@@ -118,7 +118,7 @@ namespace rocsparse
 #define LAUNCH_BSRSV_UPPER_SHARED(bsize, wfsize, dim, arch)           \
     RETURN_IF_HIPLAUNCHKERNELGGL_ERROR(                               \
         (bsrsv_upper_shared<bsize, wfsize, dim, arch>),               \
-        dim3((wfsize * mb - 1) / bsize + 1),                          \
+        dim3((static_cast<int64_t>(wfsize) * mb - 1) / bsize + 1),    \
         dim3(bsize),                                                  \
         0,                                                            \
         stream,                                                       \
@@ -170,7 +170,7 @@ namespace rocsparse
 #define LAUNCH_BSRSV_LOWER_GENERAL(bsize, wfsize, arch)               \
     RETURN_IF_HIPLAUNCHKERNELGGL_ERROR(                               \
         (bsrsv_lower_general<bsize, wfsize, arch>),                   \
-        dim3((wfsize * mb - 1) / bsize + 1),                          \
+        dim3((static_cast<int64_t>(wfsize) * mb - 1) / bsize + 1),    \
         dim3(bsize),                                                  \
         0,                                                            \
         stream,                                                       \
@@ -193,7 +193,7 @@ namespace rocsparse
 #define LAUNCH_BSRSV_UPPER_GENERAL(bsize, wfsize, arch)               \
     RETURN_IF_HIPLAUNCHKERNELGGL_ERROR(                               \
         (bsrsv_upper_general<bsize, wfsize, arch>),                   \
-        dim3((wfsize * mb - 1) / bsize + 1),                          \
+        dim3((static_cast<int64_t>(wfsize) * mb - 1) / bsize + 1),    \
         dim3(bsize),                                                  \
         0,                                                            \
         stream,                                                       \
